refactor: Modulariza el index y actualiza a módulos ES6
Se cambió de CommonJS a ES6 para modernizar la base del código y mejorar la mantenibilidad.
