+++ WhatsAppHub/src/app/middlewares/errorHandler.js
@@ -1,0 +1,7 @@
+export default (err, req, res, next) => {
+  console.error('Error:', err);
+  res.status(500).json({
+    error: 'Error interno del servidor',
+    message: err.message
+  });
+};
