+++ WhatsAppHub/src/app/routes/health.routes.js
@@ -1,0 +1,8 @@
+import { Router } from 'express';
+import { getHealthStatus } from '../controllers/health.controller.js';
+
+const router = Router();
+
+router.get('/health', getHealthStatus);
+
+export default router;
