+++ WhatsAppHub/src/assets/example-card-preview.png
