+++ WhatsAppHub/src/assets/extensions/ExampleCard.jsx
