+++ WhatsAppHub/src/assets/extensions/example-card.json
