+++ WhatsAppHub/src/assets/extensions/package.json
