+++ package.json
@@ -2,6 +2,7 @@
   "name": "whatsapphub",
   "version": "1.0.0",
   "main": "index.js",
+  "type": "module",
   "scripts": {
     "start": "NODE_ENV=production node index.js",
     "dev": "nodemon index.js",
